backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
Pillow
pytesseract
pydantic-settings
google-generativeai
python-dotenv

